apps/backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3
google-api-python-client==2.129.0
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
langgraph==0.0.50
langchain-core==0.2.5
openai==1.35.7
httpx==0.27.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pypdf==4.2.0
pytest==8.2.2
python-multipart==0.0.9
python-dateutil==2.9.0.post0
pgvector==0.3.6
google-generativeai>=0.8.0
stripe>=10.0.0
PyJWT>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
icalendar>=5.0.0
